move authentication logic from component DockGuest to an event used in multiple layouts
