+++ frontend/src/components/Dock/DockGuest.tsx
@@ -1,51 +1,15 @@
-import { useEffect } from "react";
 import { Dock } from "./Dock";
-import authService from "../../services/auth";
-
-import { useSessionContext } from "../../context/SessionContext";
-
 import Link from "next/link";
 import { IconButton, Tooltip } from "@mui/material";
 import Image from "next/image";
-
 import styles from "../../styles/Home.module.css";
 import FTLogo from "../../public/42logo.png";
-
-import Cookies from "js-cookie";
-
-import { errorHandler } from "../../errors/errorHandler";
-
-import { useErrorContext } from "../../context/ErrorContext";
-import { useSocketContext } from "../../context/SocketContext";
-
 import { SecondFactorLogin } from "../Alerts/SecondFactorLogin";
 import getConfig from "next/config";
 
 const { publicRuntimeConfig } = getConfig();
 
 export function DockGuest() {
-  const sessionContext = useSessionContext();
-  const errorContext = useErrorContext();
-  const socketContext = useSocketContext();
-
-  const authenticate = () => {
-    if (Cookies.get(publicRuntimeConfig.ACCESSTOKEN_COOKIE_NAME)) {
-      authService
-        .getLoggedInUser()
-        .then((userSelf) => {
-          sessionContext.login?.(userSelf);
-          socketContext.socket.emit("user:new", userSelf.login42);
-        })
-        .catch((error) => {
-          errorContext.newError?.(errorHandler(error, sessionContext));
-        });
-    }
-  };
-
-  useEffect(() => {
-    authenticate();
-  }, [sessionContext]);
-
   return (
     <>
       <Dock>
