+++ frontend/src/layouts/indexLayout.tsx
@@ -1,9 +1,21 @@
-import React from "react";
+import React, { useEffect } from "react";
 import styles from "../styles/Home.module.css";
 import Link from "next/link";
 import { DockSelector } from "../components/Dock/DockSelector";
+import { useSessionContext } from "../context/SessionContext";
+import { useErrorContext } from "../context/ErrorContext";
+import { useSocketContext } from "../context/SocketContext";
+import { authenticate } from "../events/authenticate";
 
 export const IndexLayout = ({ children }: { children: React.ReactNode }) => {
+  const sessionContext = useSessionContext();
+  const errorContext = useErrorContext();
+  const socketContext = useSocketContext();
+
+  useEffect(() => {
+    authenticate(errorContext, socketContext, sessionContext);
+  }, [sessionContext]);
+
   return (
     <>
       {children}
